+++ save-agent/src/linuxX64Main/resources/agent.properties
@@ -7,3 +7,4 @@ retry.attempts=5
 retry.initialRetryMillis=2000
 debug=true
 cliCommand=echo Provide CLI command here
+testSuitesDir=.
